requirements.txt
streamlit>=1.36
pandas>=2.1
numpy>=1.26
scipy>=1.11
statsmodels>=0.14
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20
openpyxl>=3.1
python-docx>=1.1
networkx>=3.2
semopy>=2.3
